requirements.txt
aiogram>=3.0.0
cerebras-cloud-sdk
python-dotenv